# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Animate Auto Tool API
  version: 1.0.0
  description: JSON contract consumed by the embedded Vue application.
servers:
  - url: /api/v1
security:
  - cookieSession: []
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }]
      responses: { "200": { $ref: "#/components/responses/Paginated" } }
    post:
      operationId: createSubscription
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SubscriptionInput" } } } }
      responses: { "201": { $ref: "#/components/responses/Success" }, "400": { $ref: "#/components/responses/Error" } }
...
  /subscriptions/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    put: { operationId: updateSubscription, requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SubscriptionInput" } } } }, responses: { "200": { $ref: "#/components/responses/Success" } } }
    delete: { operationId: deleteSubscription, responses: { "200": { $ref: "#/components/responses/Success" } } }
...
  /subscriptions/{id}/run:
    post: { operationId: runSubscription, parameters: [{ $ref: "#/components/parameters/Id" }], responses: { "202": { $ref: "#/components/responses/TaskAccepted" } } }
...
components:
  securitySchemes:
    cookieSession: { type: apiKey, in: cookie, name: animate_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 100 } }
...
  responses:
    Success: { description: Successful response, content: { application/json: { schema: { $ref: "#/components/schemas/Envelope" } } } }
    Paginated: { description: Paginated successful response, content: { application/json: { schema: { $ref: "#/components/schemas/PaginatedEnvelope" } } } }
    TaskAccepted: { description: Background task accepted, content: { application/json: { schema: { $ref: "#/components/schemas/TaskEnvelope" } } } }
    Error: { description: Error response, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  schemas:
...
    SubscriptionInput:
      type: object
      required: [title, rss_url]
      properties:
        title: { type: string }
        rss_url: { type: string, format: uri }
        backup_rss_url: { type: string, format: uri }
        filter_rule: { type: string }
        exclude_rule: { type: string }
        expected_episodes: { type: integer, minimum: 0 }
        allow_multi_subgroup: { type: boolean }
        auto_disable_on_done: { type: boolean }
        stale_after_hours: { type: integer, minimum: 1 }
    Envelope:
      type: object
      required: [data]
      properties: { data: {}, message: { type: string }, meta: { type: object, additionalProperties: true } }
    PaginatedEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [meta]
          properties:
            meta:
              type: object
              required: [page, page_size, total]
              properties: { page: { type: integer }, page_size: { type: integer }, total: { type: integer } }
    TaskEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [task_id, status]
              properties: { task_id: { type: string }, status: { type: string, const: running } }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties: { code: { type: string }, message: { type: string } }